gsd: Add basic Prm data builder
This builder type can be used to build up the Prm data bytes at some
point.

+++ gsd/src/parser.rs
@@ -95,9 +95,9 @@ pub fn parse(file: &std::path::Path, source: &str) -> crate::GenericStationDescr
                     _ => unreachable!(),
                 };
 
-                let default_value = parse_number(content.next().unwrap());
-                let min_value = parse_number(content.next().unwrap());
-                let max_value = parse_number(content.next().unwrap());
+                let default_value = parse_number(content.next().unwrap()) as i64;
+                let min_value = parse_number(content.next().unwrap()) as i64;
+                let max_value = parse_number(content.next().unwrap()) as i64;
 
                 let mut text_ref = None;
                 for data_setting in content {
